% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coef}
\usage{
\method{coef}{linreg}(linreg_object)
}
\arguments{
\item{linreg_object}{e}
}
\value{
coeffients
}
\description{
Coef
}
\examples{
data(iris)
s<-linreg(formula=Petal.Length~Species, data=iris)
coef(s)
}
