% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linreg}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{formula object}

\item{data}{dataframe}
}
\value{
object of type linreg
}
\description{
Linreg
}
\examples{
data(iris)
s<-linreg(formula=Petal.Length~Species, data=iris)
}
