% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{pred}
\usage{
pred(object)
}
\arguments{
\item{object}{\itemize{
\item object of type linreg
}}
}
\value{
fitted values
}
\description{
pred
}
\examples{
data(iris)
s<-linreg(formula=Petal.Length~Species, data=iris)
pred(s)
}
