% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals}
\usage{
\method{residuals}{linreg}(linreg_object)
}
\arguments{
\item{linreg_object}{object of type linreg}
}
\value{
residuals
}
\description{
Residuals
}
\examples{
data(iris)
s<-linreg(formula=Petal.Length~Species, data=iris)
residuals(s)
}
