% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary}
\usage{
\method{summary}{linreg}(linreg_object)
}
\arguments{
\item{linreg_object}{object of type linreg}
}
\value{
prints output
}
\description{
Summary
}
\examples{
s<-linreg(formula=Petal.Length~Sepal.Width+Sepal.Length, data=iris)
summary(s)
}
