% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coef}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{e}

\item{...}{arguments to be passed to methods}
}
\value{
coeffients
}
\description{
Coef
}
\examples{
data(iris)
s<-linreg(formula=Petal.Length~Species, data=iris)
coef(s)
}
