% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{arguments to be passed to methods}
}
\value{
ggplot object
}
\description{
Plot
}
\examples{
data(iris)
linres <- linreg(formula=Sepal.Length ~ Petal.Length + Petal.Width, data=iris)
plot(linres)
}
