% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print}
\usage{
\method{print}{linreg}(linreg_object)
}
\arguments{
\item{linreg_object}{object of type linreg}
}
\value{
prints output
}
\description{
Print
}
\examples{
data(iris)
s<-linreg(formula=Petal.Length~Species, data=iris)
print(s)
}
